tools: Break up pygments import script
